+++ CHANGELOG.md
@@ -16,6 +16,7 @@ For a full diff see [`3.1.0...3.0.0`][2.1.0...3.0.0].
 
 - Added `Specification::closure()` ([#56]), by [@localheinz]
 - Added `Specification::never()` ([#57]), by [@localheinz]
+- Added `Specification::always()` ([#58]), by [@localheinz]
 
 ## [`3.0.0`][3.0.0]
 
@@ -89,5 +90,6 @@ For a full diff see [`a5ba52c...1.0.0`][a5ba52c...1.0.0].
 [#53]: https://github.com/ergebnis/json-pointer/pull/53
 [#56]: https://github.com/ergebnis/json-pointer/pull/56
 [#57]: https://github.com/ergebnis/json-pointer/pull/57
+[#58]: https://github.com/ergebnis/json-pointer/pull/58
 
 [@localheinz]: https://github.com/localheinz